Moved into seperate file.
