+++ numba_dpex/core/passes/parfor.py
@@ -3487,93 +3487,6 @@ class ParforPass(ParforPassStates):
         return _mk_parfor_loops(self.typemap, size_vars, scope, loc)
 
 
-class ParforCFDPass(ParforPassStates):
-
-    """ParforCFDPass class is responsible for enforce CFD"""
-
-    def run(self):
-        # Get parfor params to calculate reductions below.
-        _, parfors = get_parfor_params(
-            self.func_ir.blocks,
-            self.options.fusion,
-            self.nested_fusion_info,
-        )
-
-        paramsNameSet = set()
-        from numba_dpex.core.exceptions import ComputeFollowsDataInferenceError
-
-        topo_order = find_topo_order(self.func_ir.blocks)
-        for label in topo_order:
-            block = self.func_ir.blocks[label]
-            for stmt in block.body:
-                if isinstance(stmt, ir.Assign) and isinstance(
-                    stmt.value, ir.Expr
-                ):
-                    lhs = stmt.target.name
-                    rhs = stmt.value
-
-                    print("----------->lhs= ", lhs)
-                    print("----------->rhs= ", rhs)
-
-                    if rhs.op == "call":
-                        for ele in rhs.list_vars():
-                            if ele.name in paramsNameSet:
-                                raise ComputeFollowsDataInferenceError
-
-                elif isinstance(stmt, Parfor):
-                    print("---->params: ", stmt.params)
-                    print("----->stmt", stmt)
-
-                    if stmt.params == None:
-                        continue
-                    outputParams = get_parfor_outputs(stmt, stmt.params)
-                    if len(outputParams) > 1:
-                        raise AssertionError
-                    intputDevice = None
-                    for para in stmt.params:
-                        if not isinstance(self.typemap[para], DpnpNdArray):
-                            continue
-                        # breakpoint()
-                        paramsNameSet.add(para)
-                        # assuming DpnpNdArray either input or output, and no intermdiate value
-                        if para not in outputParams:
-                            if intputDevice == None:
-                                intputDevice = self.typemap[para].device
-                            if intputDevice != self.typemap[para].device:
-                                raise ComputeFollowsDataInferenceError
-
-                    # only update device
-                    # FIXME: sycl_queue will be updated later
-                    if intputDevice == None:
-                        raise AssertionError
-
-                    if not isinstance(
-                        self.typemap[outputParams[0]], DpnpNdArray
-                    ):
-                        continue
-                    if self.typemap[outputParams[0]].device == intputDevice:
-                        continue
-                    # now update typemap
-                    self.typemap[outputParams[0]].device = intputDevice
-
-        # # check input and output arrays in parfor are same type.
-        # for parfor in parfors:
-        #     locType = None
-        #     for para in parfor.params:
-        #         if not isinstance(self.typemap[para], DpnpNdArray):
-        #             continue
-        #         if locType == None:
-        #             locType = self.typemap[para]
-        #         if self.typemap[para] != locType:
-        #             raise AssertionError
-
-        #     from .parfor_lowering_pass import _lower_parfor_gufunc
-
-        #     if locType and isinstance(locType, DpnpNdArray):
-        #         parfor.lowerer = _lower_parfor_gufunc
-        #         # breakpoint()
-
-
 class ParforFusionPass(ParforPassStates):
 
     """ParforFusionPass class is responsible for fusing parfors"""
